app/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
apache-airflow==2.8.4
pandas>=2.2.3
sqlalchemy>=1.4.41
psycopg2-binary>=2.9.6
boto3>=1.35.72
numpy>=2.0.2
email-validator==1.3.1
